feat: add confirmation dialog before deleting leave/leave req

+++ src/app/pages/new-vacation-page/new-vacation-page.component.ts
@@ -114,6 +114,8 @@ export class NewVacationPageComponent implements OnInit {
 
 
   deleteRequest(id: number){
+    const isDelete = confirm("Are you sure want to delete this leave request?");
+    if (!isDelete) return;
     this.masterSrc.deleteLeaveRequestId(id).subscribe((Res: APIResponse)=>{
       this.getGriData();
     })
